backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16,<3.17
django-cors-headers>=4.7,<4.8
django-celery-beat>=2.8,<2.9
celery>=5.5,<5.6
redis>=5.2,<5.3
psycopg[binary]>=3.2,<3.3
dj-database-url>=2.3,<2.4
whitenoise>=6.9,<7.0
gunicorn>=23.0,<24.0
pytest>=8.3,<8.4
pytest-django>=4.10,<4.11
